### 2019-06-13 #1503 @whedon
[PRE REVIEW]: fitODBOD: An R Package to Model Binomial Outcome Data using Binomial Mixture and Alternate Binomial Distributions.
**Submitting author:** @Amalan-ConStat (<a href="http://orcid.org/0000-0002-0643-9052">Amalan Mahendran</a>)
**Repository:** <a href="https://github.com/Amalan-ConStat/R-fitODBOD" target ="_blank">https://github.com/Amalan-ConStat/R-fitODBOD</a>
**Version:** v1.4.0
