### 2019-06-13 #1503 @arfon
:wave: @csoneson - the submitting author suggested you as the handling editor. Is this a paper you could edit for JOSS?
