### 2019-06-13 #1503 @Amalan-ConStat
Hello @arfon and @csoneson , Thank you for assigning an editor as swiftly as possible. I hope to be at your disposal when suggestions arise regarding my paper. My apologies but I have no suggestions for suitable reviewers for my R package submission.
