### 2019-06-13 #1503 @csoneson
👋@osorensen, @nhejazi, @jjharden - would you be interested in reviewing this submission for JOSS?
