### 2019-06-13 #1503 @csoneson
@Amalan-ConStat - while we wait for the invited reviewers to respond, there are a couple of things we can fix:
- I made a pull request to your repository with some suggestions regarding grammar and wording. Please have a look and merge if you agree
- Your paper is a bit on the long side for JOSS standards. In particular, the example code describing how to use the package would fit better in the documentation
- Please remove titles (Prof.) from the author list
- Please check the references - in a few cases you need case protection (forcing capitalization by enclosing text in {}). Also please add a DOI to Grassia (1977), and remove "Retrieved from" before the DOIs of the references where it appears, for consistency.
