### 2019-06-13 #1503 @csoneson
@jjharden, @osorensen - please head over to #1505 for the review!
