+++ R/AddBin.R
@@ -433,7 +433,7 @@ NegLLAddBin<-function(x,freq,p,alpha)
 #' No.D.D=0:7         #assigning the random variables
 #' Obs.fre.1=c(47,54,43,40,40,41,39,95)     #assigning the corresponding frequencies
 #'
-#' #dontrun{}
+#' \dontrun{
 #' #estimating the probability value and alpha value
 #' results<-suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1))
 #'
@@ -442,7 +442,7 @@ NegLLAddBin<-function(x,freq,p,alpha)
 #'
 #' #extracting the estimated parameters
 #' coef(results)
-#'
+#'          }
 #'
 #' @export
 EstMLEAddBin<-function(x,freq)
@@ -663,8 +663,7 @@ coef.mlAB<-function(object,...)
 #' Obs.fre.1=c(47,54,43,40,40,41,39,95)            #assigning the corresponding the frequencies
 #'
 #' #assigning the estimated probability value
-#'
-#' #dontrun{}
+#' \dontrun{
 #' paddbin=suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1)$p)
 #'
 #' #assigning the estimated alpha value
@@ -679,7 +678,7 @@ coef.mlAB<-function(object,...)
 #'
 #' #extract fitted values
 #' fitted(results)
-#'
+#'        }
 #'
 #' @export
 fitAddBin<-function(x,obs.freq,p,alpha)
