+++ R/Triangle.R
@@ -860,7 +860,7 @@ NegLLTriBin<-function(x,freq,mode)
 #' No.D.D=0:7   #assigning the random variables
 #' Obs.fre.1=c(47,54,43,40,40,41,39,95)   #assigning the corresponding frequencies
 #'
-#' #dontrun{}
+#' \dontrun{
 #' results<-EstMLETriBin(No.D.D,Obs.fre.1)    #estimating the mode value and extracting the mode value
 #'
 #' # extract the mode value and summary
@@ -868,7 +868,7 @@ NegLLTriBin<-function(x,freq,mode)
 #' summary(results)
 #'
 #' AIC(results)  #show the AIC value
-#'
+#'         }
 #'
 #' @export
 EstMLETriBin<-function(x,freq)
