### 2019-06-14 4f4ecec
man files updated after adding dontrun
