+++ man/fitAddBin.Rd
@@ -63,8 +63,7 @@ No.D.D=0:7         #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)            #assigning the corresponding the frequencies
 
 #assigning the estimated probability value
-
-#dontrun{}
+\dontrun{
 paddbin=suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1)$p)
 
 #assigning the estimated alpha value
@@ -79,7 +78,7 @@ AIC(results)
 
 #extract fitted values
 fitted(results)
-
+       }
 
 }
 \references{
