+++ man/NegLLKumBin.Rd
@@ -39,9 +39,9 @@ messages will be provided to go further.
 No.D.D=0:7          #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)  #assigning the corresponding frequencies
 
-#dontrun{}
+\dontrun{
 NegLLKumBin(No.D.D,Obs.fre.1,1.3,4.4) #acquiring the negative log likelihood value
-
+        }
 
 }
 \references{
