+++ man/EstMLEAddBin.Rd
@@ -43,7 +43,7 @@ necessary error messages will be provided to go further.
 No.D.D=0:7         #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)     #assigning the corresponding frequencies
 
-#dontrun{}
+\dontrun{
 #estimating the probability value and alpha value
 results<-suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1))
 
@@ -52,7 +52,7 @@ summary(results)
 
 #extracting the estimated parameters
 coef(results)
-
+         }
 
 }
 \references{
