+++ man/fitKumBin.Rd
@@ -73,7 +73,7 @@ No.D.D=0:7   #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)   #assigning the corresponding frequencies
 
 #estimating the parameters using maximum log likelihood value and assigning it
-#dontrun{}
+\dontrun{
 parameters=suppressWarnings(bbmle::mle2(EstMLEKumBin,start = list(a=10.1,b=1.1,it=10000),
           data = list(x=No.D.D,freq=Obs.fre.1)))
 
@@ -91,7 +91,7 @@ fitted(results)
 
 #extracting the residuals
 residuals(results)
-
+         }
 
 }
 \references{
