+++ appveyor.yml
@@ -13,6 +13,8 @@ install:
 cache:
   - C:\RLibrary
 
+USE_RTOOLS: true
+
 environment:
   NOT_CRAN: true
   # env vars that may need to be set, at least temporarily, from time to time
