+++ appveyor.yml
@@ -13,9 +13,8 @@ install:
 cache:
   - C:\RLibrary
 
-USE_RTOOLS: true
-
 environment:
+  USE_RTOOLS: true
   NOT_CRAN: true
   # env vars that may need to be set, at least temporarily, from time to time
   # see https://github.com/krlmlr/r-appveyor#readme for details
