### 2019-06-14 95865ac
updated references by changin title, doi added and changes to the title of author
