+++ JOSSpaper/Ref.bib
@@ -1,6 +1,6 @@
 @book{collett1991,
   author={Collett, D},
-  title={Modelling Binary Data},
+  title={{Modelling Binary Data}},
   isbn={9780412387906},
   lccn={lc91023845},
   series={Chapman \& Hall/CRC Texts in Statistical Science},
@@ -17,23 +17,23 @@
   number={2},
   pages={125--148},
   publisher={{Wiley Online Library}},
-  url={https://doi.org/10.1111/j.1467-842X.1988.tb00844.x}
+  doi={10.1111/j.1467-842X.1988.tb00844.x}
 }
 
 @article{Cox1983,
-author = {Cox, D. R.},
-title = {{Some Remarks on Overdispersion}},
-journal = {Biometrika},
-volume = {70},
-number = {1},
-pages = {269--274},
-year = {1983},
-URL={https://doi.org/10.1093/biomet/70.1.269},
+  author = {Cox, D. R.},
+  title = {{Some Remarks on Overdispersion}},
+  journal = {Biometrika},
+  volume = {70},
+  number = {1},
+  pages = {269--274},
+  year = {1983},
+  doi={10.1093/biomet/70.1.269},
 }
 
 @article{Horsnell,
  ISSN = {00359238},
- url = {https://doi.org/10.2307/2342822},
+ doi = {10.2307/2342822},
  author = {G. Horsnell},
  journal = {Journal of the Royal Statistical Society. Series A (General)},
  number = {2},
@@ -46,12 +46,12 @@ URL={https://doi.org/10.1093/biomet/70.1.269},
 
 @article{Kemp1975,
 author={Kemp, Adrienne W.and Kemp, C. D.},
-title={Models for Gaussian Hypergeometric Distributions},
+title={{Models for Gaussian Hypergeometric Distributions}},
 year={1975},
 publisher={Springer Netherlands},
 pages={31--40},
 isbn={978-94-010-1842-5},
-url={https://doi.org/10.1007/978-94-010-1842-5_4}
+doi={10.1007/978-94-010-1842-5_4}
 }
 
 @article{lemmens1988,
@@ -63,7 +63,7 @@ url={https://doi.org/10.1007/978-94-010-1842-5_4}
   pages={131--135},
   year={1988},
   publisher={Rutgers University Piscataway, NJ},
-  url= { https://doi.org/10.15288/jsa.1988.49.131}
+  doi= {10.15288/jsa.1988.49.131}
 }
 
 @article{Manoj2013,
@@ -74,7 +74,7 @@ month = {Feb},
 number = {2},
 pages = {24--41},
 title = {{The McDonald Generalized Beta-Binomial Distribution: A New Binomial Mixture Distribution and Simulation Based Comparison with Its Nested Distributions in Handling Overdispersion}},
-url = {https://doi.org/10.5539/ijsp.v2n2p24},
+doi = {10.5539/ijsp.v2n2p24},
 volume = {2},
 year = {2013}
 }
@@ -82,7 +82,7 @@ year = {2013}
 @article{Rodriguez-Avi2007,
 author = {Rodriguez-Avi, J. and Conde-Sanchez, A. and Saez-Castillo, A. J. and Olmo-Jiminez, M. J.},
 issn = {00359254},
-URL={https://doi.org/10.1111/j.1467-9876.2007.00564.x},
+doi={10.1111/j.1467-9876.2007.00564.x},
 journal = {Journal of the Royal Statistical Society. Series C: Applied Statistics},
 number = {1},
 pages = {51--61},
@@ -103,13 +103,13 @@ year = {2007}
 
 @article{Karlis2008,
 author={Karlis, Dimitris and Xekalaki, Evdokia},
-title={The Polygonal Distribution},
[…]
