### 2019-06-14 349fc36
as requested the code has been removed and clear instructions are set to follow the link for a successful use of the package
