### 2019-06-14 #1505 @Amalan-ConStat
Hello reviewers, Thank you for the support. As requested I have made necessary changes for the article and merged it as well. Further there is no issue in continuous integration as there will not be a timeout.
