+++ README.md
@@ -29,6 +29,8 @@ status](https://ci.appveyor.com/api/projects/status/github/Amalan-ConStat/R-fitO
 [![Codecov test
 coverage](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD/branch/master/graph/badge.svg)](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD?branch=master)
 
+[![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
+
 # fitODBOD <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
 ## How to engage with “fitODBOD” the first time ?
