### 2019-06-14 #1505 @jjharden
I have reviewed the software and paper. Here are the issues that I see.

1. Installation instructions: I did not see a list of dependencies.
2. Example usage: There are plenty of examples, in the documentation. However, the example given on GitHub (“Just an example”) seems much simpler than what is described in the section titled “Modelling” in the paper. It would be helpful to see an example that describes demonstrates more of the software’s capabilities.
3. Automated tests: I did not see any description of automated tests or manual steps to verify the functions in the software.
3. References: Not all references have DOIs listed. Also, some have JSTOR and Google book links, which I would not recommend using.
