### 2019-06-14 b0f2eb7
resolved doi issues in Ref and main dependcies section added
