+++ JOSSpaper/Ref.bib
@@ -5,7 +5,8 @@
   lccn={lc91023845},
   series={Chapman \& Hall/CRC Texts in Statistical Science},
   year={1991},
-  publisher={Chapman \& Hall}
+  publisher={Chapman \& Hall},
+  doi={10.1002/pst.100}
 }
 
 @article{anderson1988,
@@ -98,7 +99,8 @@ year = {2007}
   volume = {27},
   number = {5},
   year = {2011},
-  pages = {511--521}
+  pages = {511--521},
+  issn = {1001-4268}
 }
 
 @article{Karlis2008,
@@ -114,7 +116,7 @@ doi={10.1007/978-0-8176-4626-4_2}
 
 @article{Correlated,
  ISSN = {0006341X, 15410420},
- URL = {http://www.jstor.org/stable/2529589},
+ doi = {10.2307/2529589},
  author = {L. L. Kupper and J. K. Haseman},
  journal = {Biometrics},
  number = {1},
@@ -174,13 +176,12 @@ year = {1977}
 
 
 @book{johnson1995,
-  title={{Continuous univariate distributions}},
+  title={{Continuous Univariate Distributions}},
   author={Johnson, N.L. and Kotz, S. and Balakrishnan, N.},
   number={v. 2},
   isbn={9780471584940},
   lccn={93045348},
   series={Wiley series in probability and mathematical statistics: Applied probability and statistics},
-  url={https://books.google.lk/books?id=0QzvAAAAMAAJ},
   year={1995},
   publisher={Wiley \& Sons}
 }
