+++ R/GHGbeta.R
@@ -51,7 +51,7 @@
 #' Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization
 #' of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.
 #'
-#' Available at : \url{http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x}
+#' Available at : \url{https://doi.org/10.1111/j.1467-9876.2007.00564.x}
 #'
 #' Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.
 #'
@@ -206,7 +206,7 @@
 #' Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization
 #' of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.
 #'
-#' Available at : \url{http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x}
+#' Available at : \url{https://doi.org/10.1111/j.1467-9876.2007.00564.x}
 #'
 #' Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.
 #'
@@ -363,7 +363,7 @@ pGHGBeta<-function(p,n,a,b,c)
 #' Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization
 #' of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.
 #'
-#' Available at : \url{http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x}
+#' Available at : \url{https://doi.org/10.1111/j.1467-9876.2007.00564.x}
 #'
 #' Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.
 #'
@@ -530,7 +530,7 @@ mazGHGBeta<-function(r,n,a,b,c)
 #' Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization
 #' of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.
 #'
-#' Available at : \url{http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x}.
+#' Available at : \url{https://doi.org/10.1111/j.1467-9876.2007.00564.x}.
 #'
 #' Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.
 #'
@@ -677,7 +677,7 @@ dGHGBB<-function(x,n,a,b,c)
 #' Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization
 #' of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.
 #'
-#' Available at : \url{http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x}
+#' Available at : \url{https://doi.org/10.1111/j.1467-9876.2007.00564.x}
 #'
 #' Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.
 #'
[…]
