+++ man/EstMLETriBin.Rd
@@ -61,7 +61,7 @@ Series A, 120:148-191.
 Karlis, D. & Xekalaki, E., 2008. The Polygonal Distribution. In Advances in Mathematical and Statistical
 Modeling. Boston: Birkhuser Boston, pp. 21-33.
 
-Available at: \url{http://dx.doi.org/10.1007/978-0-8176-4626-4_2}.
+Available at: \url{https://doi.org/10.1007/978-0-8176-4626-4_2}.
 
 Okagbue, H. et al., 2014. Using the Average of the Extreme Values of a Triangular Distribution for a
 Transformation, and Its Approximant via the Continuous Uniform Distribution. British Journal of Mathematics
