### 2019-06-15 #1505 @osorensen
I cannot find a release in the GitHub repository. Please create a release [here](https://github.com/Amalan-ConStat/R-fitODBOD/releases) corresponding to v1.4.0.
