### 2019-06-15 Amalan-ConStat/R-fitODBOD#3 @osorensen
Comments on manuscript
Here are my comments on the [manuscript submitted to JOSS](https://github.com/openjournals/joss-reviews/issues/1505). You might have to open it with Adobe Acobat Reader (freely available at https://get.adobe.com/reader/) for the comments to show correctly.

Click this link to open the reviewed manuscript: [10.21105.joss.01505.pdf](https://github.com/Amalan-ConStat/R-fitODBOD/files/3293085/10.21105.joss.01505.pdf)
