### 2019-06-15 Amalan-ConStat/R-fitODBOD#3 @Amalan-ConStat
I will address the comments soon and will update the repository. Thank you.
