### 2019-06-15 Amalan-ConStat/R-fitODBOD#4 @osorensen
EstMLE methods
This is part of my [review of your JOSS submission](https://github.com/openjournals/joss-reviews/issues/1505).

Having to use `suppressWarnings()` in order to use `bbmle::mle2` does not seem like best practice to me.

- First, why does `bbmle::mle2` give warnings in the first place, and are they harmless?
- Second, if `suppressWarnings()` is absolutely necessary, could you hide it with a wrapper function?

Regarding my second point, I am thinking about the following call:

```r
parameters=suppressWarnings(bbmle::mle2(EstMLEGrassiaIIBin,start = list(a=0.1,b=0.1),
data = list(x=No.D.D,freq=Obs.fre.1)))
```

You could define the function `EstMLEGrassiaIIBin` such that it calls `bbmle::mle2` with the right parameters, abstracting away this complexity for the user. I have here used `.EstMLEGrassiaIIBin` to denote the current `EstMLEGrassiaIIBin`. This would now be a non-exported worker function which the user does not interact with.

```r
# Suggested new function definition:
EstMLEGrassiaIIBin <- function(x, freq, a_start = 0.1, b = 0.1){
  parameters=suppressWarnings(bbmle::mle2(.EstMLEGrassiaIIBin,start = list(a=a_start,b=b_start),
data = list(x=x,freq=freq)))

  return(parameters)
}

```

The user could then just write, in the same manner as for the MGF method.

```r
EstMLEGrassiaIIBin(x, freq)
```

The same applies to the other EstMLE method.
