### 2019-06-15 Amalan-ConStat/R-fitODBOD#4 @Amalan-ConStat
@osorensen ,Thank you, First the warnings are harmless and they do occur when the Log likelihood values are NAN for some non convergent iterations of the optimization of given parameters. This does not harm the analysis at any point. Further, yes it is possible to hide it with a wrapper function, which i am able to do and submit.
