### 2019-06-15 e7db54c
updated for Binomial, Grassia ref capitalization, some spaces and adding have
