+++ JOSSpaper/paper.md
@@ -39,7 +39,7 @@ fitting in this package. Alternate Binomial Distributions such as
 Additive Binomial (AddBin), Beta-Correlated Binomial (BetaCorrBin), COM
 Poisson Binomial (COMPBin), Correlated Binomial (CorrBin), Lovinson
 Multiplicative Binomial (LMBin) and Multiplicative Binomial (MultiBin)
-distributions are used as well, replacing the traditional Binomial
+distributions are used as well, replacing the traditional binomial
 distribution. Further, Probability Mass Function (PMF), Cumulative
 Probability Mass Function (CPMF), Negative Log Likelihood,
 Over-dispersion and parameter estimation (shape and distribution
@@ -48,24 +48,24 @@ distinct parameters) can be explored for each fitted model with the
 
 # Introduction
 
-Statistical methods are widely used for research in most disciplines. There is a focus towards fitting
-distributions to given data since the distributions of data depends on
-the method of data collection. For example, consider a Binomial
+Statistical methods are widely used for research in most disciplines. There is a 
+focus towards fitting distributions to given data since the distributions of 
+data depends on the method of data collection. For example, consider a binomial
 experiment where a fair coin is being tossed *n* times. Let the
 event of landing heads-up be defined as the success of probability *p*.
 Then, the number of heads out of *n* tosses is considered to be a single
-Binomial variable, *Y*. Also if similar Binomial experiments occur in
+binomial variable, *Y*. Also if similar binomial experiments occur in
 *N* different clusters, a collection of *Y<sub>1<sub>*,
 *Y<sub>2<sub>*, *Y<sub>3<sub>*, ..., *Y<sub>N<sub>* would form the BOD.
 Such data are frequently mentioned in fields of toxicology, biology,
 clinical medicine, epidemiology and many more. One may attempt to fit
-the BOD using the traditional Binomial distribution, as it is
+the BOD using the traditional binomial distribution, as it is
 characterized using the number of identical trials *n* and the
 probability of success parameter *p*. The parameter *p* (*p* $\in$ [0, 1]) 
 is usually assumed to be a constant from trial to trial and
 the trials are independent. In many empirical situations, it has been
 frequently observed that the actual observed variance of the BOD is
-greater than the assumed theoretical Binomial variance. This outcome is
+greater than the assumed theoretical binomial variance. This outcome is
 typically known as "over-dispersion" [@Cox1983; @anderson1988].
 Over-dispersion in BOD can occur either with a probability of success
 parameter *p* varying from trial to trial or if there is a correlation
@@ -73,31 +73,31 @@ among binary trials. However, @collett1991 argued that the above two
 cases of over-dispersion are frequently the same.
 
 New distributions emerged to fit the BOD replacing the traditional
-Binomial distribution. @Xiaohu2011 has developed the
-Kumaraswamy Binomial distribution, @Rodriguez-Avi2007 has
+binomial distribution. @Xiaohu2011 have developed the
+Kumaraswamy Binomial distribution, @Rodriguez-Avi2007 have
 constructed the Gaussian Hypergeometric Generalized Beta-Binomial
 distribution, @Karlis2008 wrote the article on the
 Triangular Binomial distribution. Also, @grassia1977 mentioned the 
 Gamma Binomial and Grassia II Binomial distributions. The Beta-Binomial 
 distribution is clearly explained in @johnson1995. Initially the concept 
-of mixing the Binomial distribution with a unit bounded continuous distribution 
+of mixing the binomial distribution with a unit bounded continuous distribution 
 was done by @Horsnell, which led to the Uniform Binomial distribution. 
[…]
