+++ JOSSpaper/Ref.bib
@@ -151,7 +151,7 @@ doi={10.1007/978-0-8176-4626-4_2}
 }
 
 @article{Multiplicative,
-  author = {S.R.Paul},
+  author = {Paul, S.R.},
   title = {{A three-parameter generalization of the Binomial Distribution}},
   journal = {Communications in Statistics - Theory and Methods},
   volume = {14},
@@ -165,7 +165,7 @@ doi={10.1007/978-0-8176-4626-4_2}
 
 @article{grassia1977,
 author = {Grassia, A.},
-title = {{ON A FAMILY OF DISTRIBUTIONS WITH ARGUMENT BETWEEN 0 AND 1 OBTAINED BY TRANSFORMATION OF THE GAMMA AND DERIVED COMPOUND DISTRIBUTIONS}},
+title = {{On a Family of Distributions with argument between 0 and 1 obtained by transformation of the Gamma and Derived Compund Distributions}},
 journal = {Australian Journal of Statistics},
 volume = {19},
 number = {2},
