### 2019-06-16 a435d85
updating EstMLE functions with suppresswarnings
