+++ man/fitGrassiaIIBin.Rd
@@ -67,8 +67,7 @@ No.D.D=0:7       #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)          #assigning the corresponding frequencies
 
 #estimating the parameters using maximum log likelihood value and assigning it
-parameters=suppressWarnings(bbmle::mle2(EstMLEGrassiaIIBin,start = list(a=0.1,b=0.1),
-data = list(x=No.D.D,freq=Obs.fre.1)))
+parameters=EstMLEGrassiaIIBin(x=No.D.D,freq=Obs.fre.1,a=0.1,b=0.1)
 
 aGIIBin=bbmle::coef(parameters)[1]         #assigning the estimated a
 bGIIBin=bbmle::coef(parameters)[2]         #assigning the estimated b
