+++ man/fitMcGBB.Rd
@@ -72,8 +72,7 @@ Obs.fre.1=c(47,54,43,40,40,41,39,95)          #assigning the corresponding frequ
 
 #estimating the parameters using maximum log likelihood value and assigning it
 \dontrun{
-parameters=suppressWarnings(bbmle::mle2(EstMLEMcGBB,start = list(a=0.1,b=0.1,c=0.2),
-data = list(x=No.D.D,freq=Obs.fre.1)))
+parameters=EstMLEMcGBB(x=No.D.D,freq=Obs.fre.1,a=0.1,b=0.1,c=3.2)
 
 aMcGBB=bbmle::coef(parameters)[1]         #assigning the estimated a
 bMcGBB=bbmle::coef(parameters)[2]         #assigning the estimated b

[…]
