### 2019-06-17 af79e17
EstMLE functions completely updated
