+++ man/fitMultiBin.Rd
@@ -64,8 +64,7 @@ No.D.D=0:7       #assigning the random variables
 Obs.fre.1=c(47,54,43,40,40,41,39,95)     #assigning the corresponding frequencies
 
 #estimating the parameters using maximum log likelihood value and assigning it
-parameters=suppressWarnings(bbmle::mle2(EstMLEMultiBin,start = list(p=0.1,theta=.3),
-          data = list(x=No.D.D,freq=Obs.fre.1)))
+parameters=EstMLEMultiBin(x=No.D.D,freq=Obs.fre.1,p=0.1,theta=.3)
 
 pMultiBin=bbmle::coef(parameters)[1]    #assigning the estimated probability value
 thetaMultiBin=bbmle::coef(parameters)[2]  #assigning the estimated theta value

[…]
