### 2019-06-17 Amalan-ConStat/R-fitODBOD#4 @osorensen
@Amalan-ConStat , thanks for the answers. In my opinion a wrapper function would make life easier for the user. In addition, `suppressWarnings` also suppresses warnings that may be useful to be aware of, and hence I do not think it is good practice to use it by default.

The following example illustrates some different calls to `bbmle::mle2`, including a case where the warning should not be ignored:

``` r
library(fitODBOD)
#> Hello, This is Amalan. For more details refer --> https://amalan-constat.github.io/R-fitODBOD/index.html

# Estimate shape parameters of Beta-Binomial Distribution
bbmle::mle2(EstMLEBetaBin, start = list(a = .1, b = .1),
            data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2))
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#> Warning in beta(a + data[i], n + b - data[i]): NaNs produced
#> Warning in beta(a, b): NaNs produced
#>
#> Call:
#> bbmle::mle2(minuslogl = EstMLEBetaBin, start = list(a = 0.1,
#>     b = 0.1), data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2))
#>
#> Coefficients:
#>         a         b
#> 0.8575483 0.7007740
#>
#> Log-likelihood: -821.39

# Suppress the warnings of the above call
suppressWarnings(
  bbmle::mle2(EstMLEBetaBin, start = list(a = .1, b = .1),
              data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2))

)
#>
#> Call:
#> bbmle::mle2(minuslogl = EstMLEBetaBin, start = list(a = 0.1,
#>     b = 0.1), data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2))
#>
#> Coefficients:
#>         a         b
#> 0.8575483 0.7007740
#>
#> Log-likelihood: -821.39

# What if the user tries to set lower bounds on a and b: This gives a warning
# message that should not be ignored
bbmle::mle2(EstMLEBetaBin, start = list(a = .1, b = .1),
            data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2),
            lower = c(0, 0))
#> Warning in optim(par = c(a = 0.1, b = 0.1), fn = function (p) : bounds can
#> only be used with method L-BFGS-B (or Brent)
#>
#> Call:
#> bbmle::mle2(minuslogl = EstMLEBetaBin, start = list(a = 0.1,
#>     b = 0.1), data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2),
#>     lower = c(0, 0))
#>
#> Coefficients:
#>         a         b
#> 0.8575367 0.7007644
#>
#> Log-likelihood: -821.39

# However, when using suppressWarnings, the user will not see the relevant
# warning about the lower bound not being used.
suppressWarnings(
  bbmle::mle2(EstMLEBetaBin, start = list(a = .1, b = .1),
              data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2),
              lower = c(0, 0))

)
#>
#> Call:
#> bbmle::mle2(minuslogl = EstMLEBetaBin, start = list(a = 0.1,
#>     b = 0.1), data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2),
#>     lower = c(0, 0))
#>
#> Coefficients:
#>         a         b
#> 0.8575367 0.7007644
#>
#> Log-likelihood: -821.39

# If method is set to L-BFGS-B, however, the lower bounds are used and there
# should be no warning
bbmle::mle2(EstMLEBetaBin, start = list(a = .1, b = .1),
            method = "L-BFGS-B",
            data = list(x = Alcohol_data$Days, freq = Alcohol_data$week2),
            lower = c(0, 0))
#>
#> Call:
#> bbmle::mle2(minuslogl = EstMLEBetaBin, start = list(a = 0.1,
#>     b = 0.1), method = "L-BFGS-B", data = list(x = Alcohol_data$Days,
#>     freq = Alcohol_data$week2), lower = c(0, 0))
#>
#> Coefficients:
#>         a         b
#> 0.8575367 0.7007644
#>
#> Log-likelihood: -821.39
```

Because of this, I would absolutely prefer if you made wrapper functions.
