### 2019-06-17 Amalan-ConStat/R-fitODBOD#4 @Amalan-ConStat
@osorensen , Thank you for the suggestion. Yes I have noticed what you are referring here. In order to resolve that
1). I have written my own wrapper function, which would target the warnings if "NaN" is mentioned. Will that be okay !, therefore other warnings will be safe.
2). Further, using "L-BFGS-B" method does work for most of the EstMLExxx functions (from the BMD), but this is not the case for ABD.

EstMLEBetaBin<-function(x,freq,a,b,...)
{
#new suppresswarning to only focus on when "NaN" is mentioned
  suppressWarnings2 <-function(expr, regex=character())
  {
    withCallingHandlers(expr, warning=function(w)
    {
      if (length(regex) == 1 && length(grep(regex, conditionMessage(w))))
      {
        invokeRestart("muffleWarning")
      }
    }                  )
  }
  suppressWarnings2(bbmle::mle2(.EstMLEBetaBin,data=list(x=x,freq=freq),
                      start = list(a=a,b=b),...),"NaN")
}

.EstMLEBetaBin<-function(x,freq,a,b)
{
  #with respective to using bbmle package function mle2 there is no need impose any restrictions
  #therefor the output is directly a single numeric value for the negative log likelihood value of
  #beta binomial distribution
  n<-max(x)
  data<-rep(x,freq)
  i<-1:sum(freq)
  term1<-sum(log(choose(n,data[i])))
  term2<-sum(log(beta(a+data[i],n+b-data[i])))
  BetaBinLL<-term1+term2-sum(freq)*log(beta(a,b))
  return(-BetaBinLL)
}
