### 2019-06-17 Amalan-ConStat/R-fitODBOD#5 @osorensen
Consistent assignment operator in examples
This concerns the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1505).

I think the examples would be more readable if consistent assignment style was used throughout. I see the following two styles being used in the examples:

```r
a=1
a<-1
```

It would help readability if you chose one style, preferably `a <- 1` with space before and after, as it helps readability. Using `<-` rather than `=` also makes it easer to distinguish function calls or assignments within lists (e.g., `list(a = 0.1, b = 0.1)`).

Here are some lines from the examples of `fitBetaBin`, where both `a=1` and `a<-1` style is used.

```r
aBetaBin=bbmle::coef(parameters)[1]  #assigning the parameter a
bBetaBin=bbmle::coef(parameters)[2]  #assigning the parameter b

#fitting when the random variable,frequencies,shape parameter values are given.
fitBetaBin(No.D.D,Obs.fre.1,aBetaBin,bBetaBin)

#estimating the parameters using moment generating function methods
results<-EstMGFBetaBin(No.D.D,Obs.fre.1)

```
