### 2019-06-17 Amalan-ConStat/R-fitODBOD#4 @osorensen
Yes, @Amalan-ConStat , I think this looks good. And the `@export` statement should then be above `EstMLEBetaBin`, and not above `.EstMLEBetaBin`, right?
