### 2019-06-17 Amalan-ConStat/R-fitODBOD#4 @Amalan-ConStat
Yes, that is correct. sorry for that mistake. Also, will there be any issues if the user uses another language.
I did use the phrase "NaNs produced" before but it does not work for the French language that is why i reduced the phrase to "NaN".
