### 2019-06-17 Amalan-ConStat/R-fitODBOD#5 @Amalan-ConStat
Sorry for the inconsistency and yes I can maintain assignment styles throughout all the examples. Thank you for pointing this out.
