### 2019-06-17 1633672
updating assignment styles with consistency
