+++ vignettes/BMDs_and_ABDs_EstMLExxxBin.R
@@ -5,7 +5,6 @@ library(ggplot2)
 library(reshape2)
 library(grid)
 library(gridExtra)
-library(bbmle)
 
 ## ----Printing data-------------------------------------------------------
 # estimating parameter for p and alpha using Additive Binomial Distribution of Chromosome data
@@ -13,104 +12,93 @@ Chromosome_data
 
 ## ----printing estimated parameters for Additive Binomial Distribution,echo=FALSE----
 # estimating parameter for p and alpha using Additive Binomial Distribution
-Est_para<-suppressWarnings(EstMLEAddBin(Chromosome_data$No.of.Asso,Chromosome_data$fre))
+Est_para<-EstMLEAddBin(Chromosome_data$No.of.Asso,Chromosome_data$fre)
 cat("Estimated parameter p for Chromosome data=",Est_para$p, "\n")
 cat("Estimated parameter alpha for Chromosome data=",Est_para$alpha,"\n")
 
 ## ----printing estimated parameter for Triangular Binomial Distribution,echo=FALSE----
 # estimating parameter for mode using Triangular Binomial Distribution of Chromosome data
-Est_para<-suppressWarnings(EstMLETriBin(Chromosome_data$No.of.Asso,Chromosome_data$fre))
+Est_para<-EstMLETriBin(Chromosome_data$No.of.Asso,Chromosome_data$fre)
 cat("Estimated parameter mode or c for Chromosome data=",Est_para$mode, "\n")
 
 ## ----printing estimated parameters for Beta-Correlated Binomial Distribution, echo=FALSE----
 #estimating parameters for Beta-Correlated Binomial Distribution of Chromosome data
-Est_para<-suppressWarnings(mle2(EstMLEBetaCorrBin,
-                           data=list(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre),
-                           start=list(cov=0.05,a=10,b=10)))
+Est_para<-EstMLEBetaCorrBin(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre,
+                            cov=0.05,a=10,b=10)
 cat("Estimated parameter covariance for Chromosome data=",coef(Est_para)[1],"\n")
 cat("Estimated parameter a for Chromosome data=",coef(Est_para)[2], "\n")
 cat("Estimated parameter b for Chromosome data=",coef(Est_para)[3], "\n")
 
 ## ----printing estimated parameters for Correlated Binomial Distribution,echo=FALSE----
 #estimating parameters for Correlated Binomial Distribution of Chromosome data
-Est_para<-suppressWarnings(mle2(EstMLECorrBin,
-                           data=list(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre),
-                           start=list(cov=0.0005,p=0.51)))
+Est_para<-EstMLECorrBin(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre,
+                        cov=0.0005,p=0.51)
 cat("Estimated parameter p for Chromosome data=",coef(Est_para)[1],"\n")
 cat("Estimated parameter covariance for Chromosome data=",coef(Est_para)[2], "\n")
 
 ## ----printing estimated parameters for COM Poisson Binomial Distribution,echo=FALSE----
 #estimating parameters for COM Poisson Binomial Distribution of Chromosome data
-Est_para<-suppressWarnings(mle2(EstMLECOMPBin,
-                           data=list(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre),
-                           start=list(v=0.0005,p=0.51)))
+Est_para<-EstMLECOMPBin(x=Chromosome_data$No.of.Asso,freq=Chromosome_data$fre,
[…]
