### 2019-06-17 cc3409f
updating the consistency in examples
