+++ R/Gbeta1.R
@@ -60,8 +60,8 @@
 #'
 #' @examples
 #' #plotting the random variables and probability values
-#' col<-rainbow(5)
-#' a<-c(.1,.2,.3,1.5,2.15)
+#' col <- rainbow(5)
+#' a <- c(.1,.2,.3,1.5,2.15)
 #' plot(0,0,main="Probability density graph",xlab="Random variable",ylab="Probability density values",
 #' xlim = c(0,1),ylim = c(0,10))
 #' for (i in 1:5)
@@ -177,8 +177,8 @@ dGBeta1<-function(p,a,b,c)
 #'
 #' @examples
 #' #plotting the random variables and probability values
-#' col<-rainbow(5)
-#' a<-c(.1,.2,.3,1.5,2.15)
+#' col <- rainbow(5)
+#' a <- c(.1,.2,.3,1.5,2.15)
 #' plot(0,0,main="Probability density graph",xlab="Random variable",ylab="Probability density values",
 #' xlim = c(0,1),ylim = c(0,10))
 #' for (i in 1:5)
@@ -303,8 +303,8 @@ pGBeta1<-function(p,a,b,c)
 #'
 #' @examples
 #' #plotting the random variables and probability values
-#' col<-rainbow(5)
-#' a<-c(.1,.2,.3,1.5,2.15)
+#' col <- rainbow(5)
+#' a <- c(.1,.2,.3,1.5,2.15)
 #' plot(0,0,main="Probability density graph",xlab="Random variable",ylab="Probability density values",
 #' xlim = c(0,1),ylim = c(0,10))
 #' for (i in 1:5)
@@ -426,8 +426,8 @@ mazGBeta1<-function(r,a,b,c)
 #'
 #' @examples
 #' #plotting the random variables and probability values
-#' col<-rainbow(5)
-#' a<-c(1,2,5,10,0.6)
+#' col <- rainbow(5)
+#' a <- c(1,2,5,10,0.6)
 #' plot(0,0,main="Mcdonald generalized beta-binomial probability function graph",
 #' xlab="Binomial random variable",ylab="Probability function values",xlim = c(0,10),ylim = c(0,0.5))
 #' for (i in 1:5)
@@ -442,8 +442,8 @@ mazGBeta1<-function(r,a,b,c)
 #' dMcGBB(0:10,10,4,2,1)$over.dis.para   #extracting the over dispersion value
 #'
 #' #plotting the random variables and cumulative probability values
-#' col<-rainbow(4)
-#' a<-c(1,2,5,10)
+#' col <- rainbow(4)
+#' a <- c(1,2,5,10)
 #' plot(0,0,main="Cumulative probability function graph",xlab="Binomial random variable",
 #' ylab="Cumulative probability function values",xlim = c(0,10),ylim = c(0,1))
 #' for (i in 1:4)
@@ -565,8 +565,8 @@ dMcGBB<-function(x,n,a,b,c)
 #'
 #' @examples
 #' #plotting the random variables and probability values
-#' col<-rainbow(5)
-#' a<-c(1,2,5,10,0.6)
+#' col <- rainbow(5)
+#' a <- c(1,2,5,10,0.6)
 #' plot(0,0,main="Mcdonald generalized beta-binomial probability function graph",
 #' xlab="Binomial random variable",ylab="Probability function values",xlim = c(0,10),ylim = c(0,0.5))
 #' for (i in 1:5)
@@ -581,8 +581,8 @@ dMcGBB<-function(x,n,a,b,c)
 #' dMcGBB(0:10,10,4,2,1)$over.dis.para   #extracting the over dispersion value
 #'
 #' #plotting the random variables and cumulative probability values
-#' col<-rainbow(4)
[…]
