### 2019-06-17 231babd
updating tests after removing warnings and replacing with errors and messages
