### 2019-06-17 737a5ad
updating consistency in vignettes
