+++ vignettes/BMDs_and_ABDs_fitxxxBin.R
@@ -1,360 +1,0 @@
-## ----setup, include=FALSE------------------------------------------------
-knitr::opts_chunk$set(echo = TRUE)
-library(fitODBOD)
-library(ggplot2)
-library(reshape2)
-library(grid)
-library(gridExtra)
-library(bbmle)
-library(tibble)
-
-## ----plotting Expected frequencies with actual frequency for Binomial Distribution, warning=FALSE----
-Alcohol_data
-
-BinRanVar<-Alcohol_data$Days
-ActFreq<-Alcohol_data$week2
-
-# Fitting Binomial Distribution
-BinFreq<-fitBin(BinRanVar,ActFreq)
-
-# printing the results of fitting Binomial distribution
-print(BinFreq)
-
-## ----plotting Expected frequencies with actual frequency for Additive Binomial Distribution, warning=FALSE----
-# Estimating and fitting Additive Binomial distribution
-Para_AddBin<-EstMLEAddBin(BinRanVar,ActFreq)
-
-# printing the coefficients and using them
-coef(Para_AddBin)
-
-AddBin_p<-Para_AddBin$p
-AddBin_alpha<-Para_AddBin$alpha
-
-# Fitting Additive Binomial Distribution
-AddBinFreq<-fitAddBin(BinRanVar, ActFreq, AddBin_p, AddBin_alpha)
-
-# printing the results of fitting Additive Binomial Distribution
-print(AddBinFreq)
-
-## ----plotting Expected frequencies with actual frequency for Beta-Correlated Binomial Distribution, warning=FALSE----
-# Estimating and fitting Beta Correlated Binoial Distribution
-Para_BetaCorrBin<-suppressWarnings(mle2(EstMLEBetaCorrBin,
-                       data=list(x=BinRanVar, freq=ActFreq),
-                       start=list(cov=0.001,a=10,b=10)))
-
-# printing the coefficients and using them
-coef(Para_BetaCorrBin)
-
-BetaCorrBin_cov<-coef(Para_BetaCorrBin)[1]
-BetaCorrBin_a<-coef(Para_BetaCorrBin)[2]
-BetaCorrBin_b<-coef(Para_BetaCorrBin)[3]
-
-# Fitting Beta-Correlated Binomial Distribution
-BetaCorrBinFreq<-suppressWarnings(fitBetaCorrBin(BinRanVar, ActFreq,BetaCorrBin_cov,BetaCorrBin_a,BetaCorrBin_b))
-
-# printing the results of fitting Beta-Correlated Binomial Distribution
-print(BetaCorrBinFreq)
-
-## ----plotting Expected frequencies with actual frequency for COM-Poisson Binomial Distribution, warning=FALSE----
-# Estimating and fitting COM Poisson Binomial Distribution
-Para_COMPBin<-suppressWarnings(mle2(EstMLECOMPBin,
-                       data=list(x=BinRanVar, freq=ActFreq),
-                       start=list(v=12.1,p=0.9)))
-
-# printing the coefficients and using them
-coef(Para_COMPBin)
-
-COMPBin_p<-coef(Para_COMPBin)[1]
-COMPBin_v<-coef(Para_COMPBin)[2]
-
-# Fitting COM-Poisson Binomial Distribution
-COMPBinFreq<-suppressWarnings(fitCOMPBin(BinRanVar, ActFreq,COMPBin_p,COMPBin_v))
-
-# printing the results of fitting COM-Poisson Binomial Distribution
-print(COMPBinFreq)
-
-## ----plotting Expected frequencies with actual frequency for Correlated Binomial Distribution, warning=FALSE----
-# Estimating and fitting Correlated Binomial Distribution
-Para_CorrBin<-suppressWarnings(mle2(EstMLECorrBin,
[…]
