+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: Modeling Over Dispersed Binomial Outcome Data Using BMD and ABD
 Description: Contains Probability Mass Functions, Cumulative Mass Functions, Negative Log Likelihood value, parameter estimation and modeling data using Binomial Mixture Distributions (BMD) (Manoj et al (2013) <doi:10.5539/ijsp.v2n2p24>) and Alternate Binomial Distributions (ABD) (Paul (1985) <doi:10.1080/03610928508828990>).
 URL: https://github.com/Amalan-ConStat/R-fitODBOD,https://amalan-constat.github.io/R-fitODBOD/index.html
 BugReports: https://github.com/Amalan-ConStat/R-fitODBOD/issues
-Version: 1.4.0
+Version: 1.4.1
 Authors@R: c(
           person("Amalan","Mahendran", email = "amalan0595@gmail.com", role = c("aut","cre")
           ,comment=c(ORCID="0000-0002-0643-9052")),
