### 2019-06-17 99b6f2f
update vignettes with consistency
