+++ vignettes/Mixing_Distributions_pxxx.Rmd
@@ -51,14 +51,15 @@ Unit Bounded Uniform distribution does not have any shape parameter. Therefore n
 Cpdf values. Although it has been plotted below
 
 ```{r Uniform Distribution plotting,fig.align='center',fig.width=9,fig.height=7}
-prob<-seq(0,1,by=0.05)
-pdfv<-pUNI(prob)
-data<-data.frame(prob,pdfv)
+prob <- seq(0,1,by=0.05)
+pdfv <- pUNI(prob)
+data <- data.frame(prob,pdfv)
 ggplot(data)+
   geom_line(aes(x=data$prob,y=data$pdfv))+
   xlab("Vector of Probabilities")+
   ylab("Cumulative Probability density values")+
   ggtitle("Cpdf values changing")+
+  ggthemes::theme_clean()+
   scale_x_continuous(breaks=seq(0,1,by=0.1))
 ```
 
@@ -68,19 +69,20 @@ Unit Bounded Triangular distribution has one parameter, which is the mode value 
 According to the below given mode parameter value restrictions it has been plotted.
 
 ```{r Triangular Distribution plotting,fig.align='center',fig.width=9,fig.height=7}
-prob<-seq(0.01,0.99,by=0.01)
-mode<-seq(0.01,0.99,by=0.01)
-output<-matrix(ncol =length(mode) ,nrow=length(prob))
+prob <- seq(0.01,0.99,by=0.01)
+mode <- seq(0.01,0.99,by=0.01)
+output <- matrix(ncol =length(mode) ,nrow=length(prob))
 for (i in 1:length(mode)) 
   {
    output[,i]<-pTRI(prob,mode[i])
   }
-data<-data.frame(prob,output)
-data<-melt(data,id.vars ="prob" )
+data <- data.frame(prob,output)
+data <- melt(data,id.vars ="prob" )
 ggplot(data,aes(prob,value,col=variable))+
   geom_line()+guides(fill=FALSE,color=FALSE)+
   xlab("Vector of Probabilities")+
   ylab("Cumulative Probability density values")+
+  ggthemes::theme_clean()+
   ggtitle("Cpdf values changing when c=seq(0.01,0.09,by0.01)")+
   scale_x_continuous(breaks=seq(0,1,by=0.1))
 ```
@@ -110,6 +112,7 @@ pBETAplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Vector of Probabilities")+
       ylab("Probability density values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -127,6 +130,7 @@ pBETAplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Vector of Probabilities")+
       ylab("Probability density values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -134,17 +138,17 @@ pBETAplot<-function(a,b,plot_title,a_seq)
 ```
 
 ```{r Beta Distribution function plotting,fig.align='center',fig.width=9,fig.height=7}
[…]
