+++ vignettes/BMDs_and_ABDs_pxxxBin.Rmd
@@ -58,7 +58,8 @@ ggplot(data)+
   geom_line(aes(x=data$brv,y=data$cpmfv))+
   xlab("Binomial Random Variable")+
   ylab("Cumulative Probability Mass Function values")+
-  ggtitle("Cpmf values changing")
+  ggtitle("Cpmf values changing")+
+  ggthemes::theme_clean()+
   scale_x_continuous(breaks=seq(0,10,by=1))
 ```
 
@@ -83,6 +84,7 @@ ggplot(data,aes(brv,value,col=variable))+
   geom_line()+guides(fill=FALSE,color=FALSE)+
   xlab("Binomial Random Variable")+
   ylab("Cumulative Probability Mass values")+
+  ggthemes::theme_clean()+
   ggtitle("Cpmf values changing for c=seq(0.02,0.98,by=0.01)")
   scale_x_continuous(breaks=seq(0,15,by=1))
 ```
@@ -111,6 +113,7 @@ pBetaBinplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -128,6 +131,7 @@ pBetaBinplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -176,6 +180,7 @@ pKumBinplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -193,6 +198,7 @@ pKumBinplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -240,6 +246,7 @@ pGHGBBplot<-function(a,b,c,plot_title,a_seq,b_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -257,6 +264,7 @@ pGHGBBplot<-function(a,b,c,plot_title,a_seq,b_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -274,6 +282,7 @@ pGHGBBplot<-function(a,b,c,plot_title,a_seq,b_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
       ylab("Cpmf values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -330,6 +339,7 @@ pMcGBBplot<-function(a,b,c,plot_title,a_seq,b_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Binomial Random Variable")+
[…]
