+++ vignettes/Mixing_Distributions_dxxx.Rmd
@@ -51,13 +51,14 @@ Uniform distribution does not involve any shape parameter and no parameter at al
 fluctuation in Pdf values. Yet plotted below
 
 ```{r Uniform Distribution,fig.align='center',fig.width=9,fig.height=7}
-prob<-seq(0.01,0.99,by=0.01)
-pdfv<-dUNI(prob)$pdf
-data<-data.frame(prob,pdfv)
+prob <- seq(0.01,0.99,by=0.01)
+pdfv <- dUNI(prob)$pdf
+data <- data.frame(prob,pdfv)
 ggplot(data)+
   geom_line(aes(x=data$prob,y=data$pdfv))+
   xlab("Vector of Probabilities")+
   ylab("Probability density values")+
+  ggthemes::theme_clean()+
   ggtitle("Pdf values changing")
 ```
 
@@ -67,19 +68,19 @@ Triangular distribution includes only one parameter, which is the mode value in-
 Below is the plot explaining how pdf values change when mode changes
 
 ```{r Triangular Distribution,fig.align='center',fig.width=9,fig.height=7}
-prob<-seq(0.01,0.99,by=0.01)
-mode<-seq(0.01,0.99,by=0.01)
-output<-matrix(ncol =length(mode) ,nrow=length(prob))
+prob <- seq(0.01,0.99,by=0.01)
+mode <- seq(0.01,0.99,by=0.01)
+output <- matrix(ncol =length(mode) ,nrow=length(prob))
 for (i in 1:length(mode)) 
   {
    output[,i]<-dTRI(prob,mode[i])$pdf
   }
-data<-data.frame(prob,output)
-data<-melt(data,id.vars ="prob" )
+data <- data.frame(prob,output)
+data <- melt(data,id.vars ="prob" )
 ggplot(data,aes(prob,value,col=variable))+
   geom_line()+guides(fill=FALSE,color=FALSE)+
   xlab("Vector of Probabilities")+
-  ylab("Probability density values")+
+  ylab("Probability density values")+ggthemes::theme_clean()+
   ggtitle("Pdf values changing when c=seq(0.01,0.09,by0.01)")
 ```
 
@@ -108,6 +109,7 @@ dBETAplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Vector of Probabilities")+
       ylab("Probability density values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -125,6 +127,7 @@ dBETAplot<-function(a,b,plot_title,a_seq)
       geom_line()+guides(fill=FALSE,color=FALSE)+
       xlab("Vector of Probabilities")+
       ylab("Probability density values")+
+      ggthemes::theme_clean()+
       ggtitle(plot_title)
       return(p1)
   }
@@ -132,17 +135,17 @@ dBETAplot<-function(a,b,plot_title,a_seq)
 ```
 
 ```{r Beta Distribution plotting, fig.align='center',fig.width=9,fig.height=7}
-b10<-dBETAplot(a=seq(1,100,by=1),b=10,plot_title="and when b=10",a_seq= T)
[…]

[…]
