### 2019-06-17 045f34f
updated tests and rbuildignore for vignettes
