+++ tests/testthat/test-fitMultiBin.R
@@ -24,7 +24,7 @@ test_that("Chi-squared approximation issues",{
 
 context("Degree of Freedom")
 test_that("Degree of freedom less than zero",{
-          expect_that(fitMultiBin(1,2,0.54,15.0003),
+          expect_that(fitMultiBin(c(0,1,2),c(11,12,12),0.54,13),
           throws_error("Degrees of freedom cannot be less than or equal to zero"))
           })
 test_that("Degree of freedom equal to zero",{
