+++ tests/testthat/test-fitTriBin.R
@@ -24,7 +24,7 @@ test_that("Chi-squared approximation issues",{
 
 context("Degree of Freedom")
 test_that("Degree of freedom less than zero",{
-          expect_that(fitTriBin(1,2,0.1),
+          expect_that(fitTriBin(c(0,1),c(110,12),0.111),
           throws_error("Degrees of freedom cannot be less than or equal to zero"))
           })
 test_that("Degree of freedom equal to zero",{
