+++ tests/testthat/test-fitLMBin.R
@@ -24,7 +24,7 @@ test_that("Chi-squared approximation issues",{
 
 context("Degree of Freedom")
 test_that("Degree of freedom less than zero",{
-          expect_that(fitLMBin(1,2,0.54,15.0003),
+          expect_that(fitLMBin(c(0,1,2),c(11,12,12),0.54,1.3),
           throws_error("Degrees of freedom cannot be less than or equal to zero"))
           })
 test_that("Degree of freedom equal to zero",{
