### 2019-06-18 491c3ab
update pkgdown and vignette
