### 2019-06-18 a62f266
updating vignettes
