+++ vignettes/Mixing_Distributions_dxxx.R
@@ -1,0 +1,419 @@
+## ----setup, include=FALSE------------------------------------------------
+knitr::opts_chunk$set(echo = TRUE)
+library(fitODBOD)
+library(ggplot2)
+library(reshape2)
+library(grid)
+library(gridExtra)
+
+## ----Uniform Distribution,fig.align='center',fig.width=9,fig.height=7----
+prob <- seq(0.01,0.99,by=0.01)
+pdfv <- dUNI(prob)$pdf
+data <- data.frame(prob,pdfv)
+ggplot(data)+
+  geom_line(aes(x=data$prob,y=data$pdfv))+
+  xlab("Vector of Probabilities")+
+  ylab("Probability density values")+
+  ggthemes::theme_clean()+
+  ggtitle("Pdf values changing")
+
+## ----Triangular Distribution,fig.align='center',fig.width=9,fig.height=7----
+prob <- seq(0.01,0.99,by=0.01)
+mode <- seq(0.01,0.99,by=0.01)
+output <- matrix(ncol =length(mode) ,nrow=length(prob))
+for (i in 1:length(mode)) 
+  {
+   output[,i]<-dTRI(prob,mode[i])$pdf
+  }
+data <- data.frame(prob,output)
+data <- melt(data,id.vars ="prob" )
+ggplot(data,aes(prob,value,col=variable))+
+  geom_line()+guides(fill=FALSE,color=FALSE)+
+  xlab("Vector of Probabilities")+
+  ylab("Probability density values")+ggthemes::theme_clean()+
+  ggtitle("Pdf values changing when c=seq(0.01,0.09,by0.01)")
+
+## ----Beta Distribution plot function, include=FALSE----------------------
+dBETAplot<-function(a,b,plot_title,a_seq)
+{
+  if(a_seq==TRUE)
+  {
+  prob<-seq(0.01,0.99,by=0.01)
+  output<-matrix(ncol=length(a),nrow=length(prob))
+    for (i in 1:length(a)) 
+    {
+    output[,i]<-dBETA(prob,a[i],b)$pdf
+    }
+  data<-data.frame(prob,output)
+  data<-melt(data,id.vars ="prob" )
+  p1<-ggplot(data,aes(prob,value,col=variable))+
+      geom_line()+guides(fill=FALSE,color=FALSE)+
+      xlab("Vector of Probabilities")+
+      ylab("Probability density values")+
+      ggthemes::theme_clean()+
+      ggtitle(plot_title)
+      return(p1)
+  }
+  if(a_seq==FALSE)
+  {
+  prob<-seq(0.01,0.99,by=0.01)
+  output<-matrix(ncol =length(b) ,nrow=length(prob))
+    for (i in 1:length(b)) 
+    {
+    output[,i]<-dBETA(prob,a,b[i])$pdf
+    }
+  data<-data.frame(prob,output)
+  data<-melt(data,id.vars ="prob" )
+  p1<-ggplot(data,aes(prob,value,col=variable))+
+      geom_line()+guides(fill=FALSE,color=FALSE)+
+      xlab("Vector of Probabilities")+
+      ylab("Probability density values")+
+      ggthemes::theme_clean()+
+      ggtitle(plot_title)
+      return(p1)
+  }
+}
+
+## ----Beta Distribution plotting, fig.align='center',fig.width=9,fig.height=7----
+ b10 <- dBETAplot(a=seq(1,100,by=1),b=10,plot_title="and when b=10",a_seq= T)
[…]
