### 2019-06-18 63e1ff5
remove vignettes but add in pkgdown
