+++ DESCRIPTION
@@ -22,4 +22,4 @@ Suggests: testthat,
       covr,
       knitr
 Encoding: UTF-8
-VignetteBuilder: knitr
+Language: en-US
