### 2019-06-20 #1505 @csoneson
@osorensen, @jjharden - thanks for your reviews! @Amalan-ConStat - please address the reviewer comments above and the issues opened in your repository, and comment back here when you are ready or if you have questions!
