### 2019-06-21 af42733
updating readme with new instructions
