+++ README.md
@@ -30,6 +30,8 @@ status](https://ci.appveyor.com/api/projects/status/github/Amalan-ConStat/R-fitO
 coverage](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD/branch/master/graph/badge.svg)](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD?branch=master)
 
 [![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
+[![cran
+checks](https://cranchecks.info/badges/summary/fitODBOD)](https://cranchecks.info/pkgs/fitODBOD)
 
 # fitODBOD <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
@@ -71,126 +73,196 @@ modeled using these Distributions
 |                                                 | 7.Gamma Binomial Distribution                                    |
 |                                                 | 8.Grassia II Binomial Distribution                               |
 
-## Just an example
+## Modelling
 
-### Modelling BOD using Beta-Binomial Distribution
+To demonstrate the process the Alcohol Consumption Data, which is the
+most commonly used dataset by the researchers to explain Over-dispersion
+will be taken {lemmens1988}. In this dataset, the number of alcohol
+consumption days in two reference weeks is separately self-reported by a
+randomly selected sample of \(399\) respondents from the Netherlands in
+\(1983\). Here, the number of days a given individual consumes alcohol
+out of seven days a week can be treated as a Binomial variable. The
+collection of all such variables from all respondents would be defined
+as “Binomial Outcome Data”.
 
-``` r
-library(fitODBOD)
+### Step 1
 
-#Looking at the BOD
-Alcohol_data
-```
-
-    ##   Days week1 week2
-    ## 1    0    47    42
-    ## 2    1    54    47
-    ## 3    2    43    54
-    ## 4    3    40    40
-    ## 5    4    40    49
-    ## 6    5    41    40
-    ## 7    6    39    43
-    ## 8    7    95    84
-
-Hypothesis to check if above data follows Binomial Distribution
-
-H<sub>0</sub> : The Data follows the Binomial Distribution.
-
-H<sub>1</sub> : The Data does not follow the Binomial Distribution.
+The Alcohol consumption data is already in the necessary format to apply
+steps \(2\) to \(5\) and hence, step \(1\) can be avoided. The steps
+\(2\) to \(5\) can be applied only if the dataset is in the form of a
+frequency table as follows.
 
 ``` r
-#Checking if the above data  of Days and week2 follows Binomial Distribution
-fitB<-fitBin(Alcohol_data$Days,Alcohol_data$week2)
-
-print(fitB)
+library("fitODBOD")   #Loading packages
+# Hello, This is Amalan. For more details refer --> https://amalan-constat.github.io/R-fitODBOD/index.html
+print(Alcohol_data)     #print the alcohol consumption data set
+#   Days week1 week2
+# 1    0    47    42
+# 2    1    54    47
+# 3    2    43    54
+# 4    3    40    40
+# 5    4    40    49
+# 6    5    41    40
+# 7    6    39    43
+# 8    7    95    84
+sum(Alcohol_data$week1) #No of respondents or N
+# [1] 399
+Alcohol_data$Days       #Binomial random variables or x 
+# [1] 0 1 2 3 4 5 6 7
 ```
 
-    ## Call: 
-    ## fitBin(x = Alcohol_data$Days, obs.freq = Alcohol_data$week2)
-    ## 
[…]
