+++ README.Rmd
@@ -4,7 +4,7 @@ output: github_document
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE,warning = FALSE,message = FALSE)
+knitr::opts_chunk$set(echo = TRUE,comment = "#",collapse = TRUE)
 ```
 
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version-last-release/fitODBOD)](https://cran.r-project.org/package=fitODBOD)
@@ -28,10 +28,12 @@ knitr::opts_chunk$set(echo = TRUE,warning = FALSE,message = FALSE)
 [![Codecov test coverage](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD/branch/master/graph/badge.svg)](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD?branch=master)
 
 [![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
+[![cran checks](https://cranchecks.info/badges/summary/fitODBOD)](https://cranchecks.info/pkgs/fitODBOD)
 
 # fitODBOD <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
 ## How to engage with "fitODBOD" the first time ?
+
 ```{r fitODBOD from GitHub or CRAN,eval=FALSE}
 ## Installing the package from GitHub
 devtools::install_github("Amalan-ConStat/R-fitODBOD")
@@ -39,6 +41,7 @@ devtools::install_github("Amalan-ConStat/R-fitODBOD")
 ## Installing the package from CRAN
 install.packages("fitODBOD")
 ```
+
 ## Key Phrases
 * BOD (Binomial Outcome Data)
 * Over Dispersion
@@ -64,109 +67,144 @@ You can understand BMD & ABD with PMF & CPMF. Further, BOD can be modeled using 
 ||7.Gamma Binomial Distribution|
 ||8.Grassia II Binomial Distribution|
 
+## Modelling
 
-## Just an example
-### Modelling BOD using Beta-Binomial Distribution
-```{r Alcohol Data}
-library(fitODBOD)
+To demonstrate the process the Alcohol Consumption Data, which is the most commonly used dataset by the researchers to explain 
+Over-dispersion will be taken {lemmens1988}. In this dataset, the number of alcohol consumption days in two reference weeks is separately
+self-reported by a randomly selected sample of $399$ respondents from the Netherlands in $1983$. Here, the number of days a given individual
+consumes alcohol out of seven days a week can be treated as a Binomial variable. The collection of all such variables from all respondents
+would be defined as "Binomial Outcome Data".
 
-#Looking at the BOD
-Alcohol_data
+### Step 1
+ 
+The Alcohol consumption data is already in the necessary format to apply steps $2$ to $5$ and hence, step $1$ can be avoided. The steps 
+$2$ to $5$ can be applied only if the dataset is in the form of a frequency table as follows.
+
+```{r ,Step 1_0}
+library("fitODBOD")   #Loading packages
+print(Alcohol_data)     #print the alcohol consumption data set
+sum(Alcohol_data$week1) #No of respondents or N
+Alcohol_data$Days       #Binomial random variables or x 
 ```
 
-Hypothesis to check if above data follows Binomial Distribution
+Suppose your dataset is not a frequency table as shown in the following dataset called `datapoints`. Then the function `BODextract` 
+can be used to prepare the appropriate format as follows.
 
-H~0~ : The Data follows the Binomial Distribution. 
-
-H~1~ : The Data does not follow the Binomial Distribution.
-
-```{r Fitting Binomial Distribution}
[…]
