### 2019-06-21 7199016
First draft of Readme changes after adding all distributions
