+++ README.Rmd
@@ -4,7 +4,7 @@ output: github_document
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE,comment = "#",collapse = TRUE)
+knitr::opts_chunk$set(echo = TRUE,comment = "#",collapse = TRUE, fig.path = "man/figures/README-")
 ```
 
 [![cran checks](https://cranchecks.info/badges/summary/fitODBOD)](https://cranchecks.info/pkgs/fitODBOD)
@@ -16,17 +16,13 @@ knitr::opts_chunk$set(echo = TRUE,comment = "#",collapse = TRUE)
 ![downloads](http://cranlogs.r-pkg.org/badges/grand-total/fitODBOD)
 ![downloads](https://cranlogs.r-pkg.org/badges/fitODBOD)
 ![downloads](http://cranlogs.r-pkg.org/badges/last-week/fitODBOD)
-[![rpackages.io rank](http://www.rpackages.io/badge/fitODBOD.svg)](http://www.rpackages.io/package/fitODBOD)
-
+[![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
 
 [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![GitHub issues](https://img.shields.io/github/issues/Amalan-ConStat/R-fitODBOD.svg?style=popout)](https://github.com/Amalan-ConStat/R-fitODBOD/issues)
 [![Travis build status](https://travis-ci.org/Amalan-ConStat/R-fitODBOD.svg?branch=master)](https://travis-ci.org/Amalan-ConStat/R-fitODBOD)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/Amalan-ConStat/R-fitODBOD?branch=master&svg=true)](https://ci.appveyor.com/project/Amalan-ConStat/R-fitODBOD)
 [![Codecov test coverage](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD/branch/master/graph/badge.svg)](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD?branch=master)
-
-[![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
-
 
 # fitODBOD <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
@@ -44,12 +40,13 @@ install.packages("fitODBOD")
 * BOD (Binomial Outcome Data)
 * Over Dispersion
 * Under Dispersion
-* BMD (Binomial Mixture Distributions)
+* FBMD (Family of Binomial Mixture Distributions)
 * ABD (Alternate Binomial Distributions)
 * PMF (Probability Mass Function)
 * CPMF (Cumulative Probability Mass Function)
 
 ## What does "fitODBOD" ?
+
 You can understand BMD & ABD with PMF & CPMF. Further, BOD can be modeled using these Distributions
 
 ## Distributions
@@ -67,25 +64,25 @@ You can understand BMD & ABD with PMF & CPMF. Further, BOD can be modeled using 
 
 ## Modelling
 
-To demonstrate the process the Alcohol Consumption Data, which is the most commonly used dataset by the researchers to explain 
-Over-dispersion will be taken {lemmens1988}. In this dataset, the number of alcohol consumption days in two reference weeks is separately
[…]
