+++ README.md
@@ -12,8 +12,7 @@ license](https://img.shields.io/github/license/Amalan-ConStat/R-fitODBOD.svg?sty
 ![downloads](http://cranlogs.r-pkg.org/badges/grand-total/fitODBOD)
 ![downloads](https://cranlogs.r-pkg.org/badges/fitODBOD)
 ![downloads](http://cranlogs.r-pkg.org/badges/last-week/fitODBOD)
-[![rpackages.io
-rank](http://www.rpackages.io/badge/fitODBOD.svg)](http://www.rpackages.io/package/fitODBOD)
+[![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
 
 [![Project Status: Active - The project has reached a stable, usable
 state and is being actively
@@ -26,8 +25,6 @@ status](https://travis-ci.org/Amalan-ConStat/R-fitODBOD.svg?branch=master)](http
 status](https://ci.appveyor.com/api/projects/status/github/Amalan-ConStat/R-fitODBOD?branch=master&svg=true)](https://ci.appveyor.com/project/Amalan-ConStat/R-fitODBOD)
 [![Codecov test
 coverage](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD/branch/master/graph/badge.svg)](https://codecov.io/gh/Amalan-ConStat/R-fitODBOD?branch=master)
-
-[![status](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4/status.svg)](http://joss.theoj.org/papers/388fc2f4d7c1e0ae83cf0de13ac038a4)
 
 # fitODBOD <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
@@ -46,7 +43,7 @@ install.packages("fitODBOD")
   - BOD (Binomial Outcome Data)
   - Over Dispersion
   - Under Dispersion
-  - BMD (Binomial Mixture Distributions)
+  - FBMD (Family of Binomial Mixture Distributions)
   - ABD (Alternate Binomial Distributions)
   - PMF (Probability Mass Function)
   - CPMF (Cumulative Probability Mass Function)
@@ -72,24 +69,24 @@ modeled using these Distributions
 ## Modelling
 
 To demonstrate the process the Alcohol Consumption Data, which is the
-most commonly used dataset by the researchers to explain Over-dispersion
-will be taken {lemmens1988}. In this dataset, the number of alcohol
-consumption days in two reference weeks is separately self-reported by a
-randomly selected sample of \(399\) respondents from the Netherlands in
-\(1983\). Here, the number of days a given individual consumes alcohol
-out of seven days a week can be treated as a Binomial variable. The
-collection of all such variables from all respondents would be defined
-as “Binomial Outcome Data”.
+most commonly used data-set by the researchers to explain
+Over-dispersion will be taken {lemmens1988}. In this data-set, the
+number of alcohol consumption days in two reference weeks is separately
+self-reported by a randomly selected sample of 399 respondents from the
+Netherlands in 1983. Here, the number of days a given individual
+consumes alcohol out of seven days a week can be treated as a Binomial
+variable. The collection of all such variables from all respondents
+would be defined as “Binomial Outcome Data”.
 
 ### Step 1
 
 The Alcohol consumption data is already in the necessary format to apply
-steps \(2\) to \(5\) and hence, step \(1\) can be avoided. The steps
[…]
