### 2019-06-23 Amalan-ConStat/R-fitODBOD#5 @Amalan-ConStat
Issues regarding to consistency have been resolved in all examples.
