### 2019-06-23 Amalan-ConStat/R-fitODBOD#4 @Amalan-ConStat
Issues regarding warning messages have been resolved. Now only errors occur based on inappropriate inputs.
