### 2019-06-23 Amalan-ConStat/R-fitODBOD#3 @Amalan-ConStat
1. List of dependencies have been added clearly.
2. Readme file in GitHub now contains a more clear view of the modeling process from step 1 to 5.
3. Yes, tests for functions are stored in the tests directory, they are unit tests using testthat package.
4. References has been changed accordingly where only 'doi' is added, but one article and book does not hold any 'doi' reference.
