### 2019-06-23 8bcb8ab
updated the paper with link and dependencies
