### 2019-06-23 #1505 @Amalan-ConStat
I have resolved all issues regarding to the repository as requested in my repository, further I have addressed all comments too.  Thank you for reminding.
