### 2019-06-23 Amalan-ConStat/R-fitODBOD#5 @osorensen
Looks good. Closing the issue.
