### 2019-06-23 Amalan-ConStat/R-fitODBOD#4 @osorensen
This looks good. Closing the issue.
