### 2019-06-23 Amalan-ConStat/R-fitODBOD#3 @osorensen
Thanks. The updated manuscript addresses all my concerns.
