### 2019-06-23 #1505 @Amalan-ConStat
I have checked the article proof, seems fine. Thank you for going through the changes swiftly.
