### 2019-06-23 #1505 @osorensen
@csoneson The updates to the software and the paper address all my concerns. I have close the issues I opened in the source repository. The software is well documented and easy to use, and implements important models which are well described in the statistical literature. Happy to accept it now.
