### 2019-06-23 #1505 @Amalan-ConStat
Thank you very much for immediate responses and clear clarification requests for the issues in the R package. How may i proceed now ?
