### 2019-06-24 #1505 @csoneson
@osorensen Thanks for your review!
@Amalan-ConStat Could you briefly summarize how you addressed @jjharden's comments above, to make it easier for him to have a second look? Regarding his point 3, it looks like most of the current unit tests are checking that the functions fail if given invalid input - could you also add some tests that confirm that they produce the correct output for given inputs?
