### 2019-06-24 705c038
updating testsin NegLL functions
