### 2019-06-24 196bd74
tests for initial ml outputs
