### 2019-06-24 8e90780
added two classes to overdispersion function
