+++ NAMESPACE
@@ -8,6 +8,8 @@ S3method(EstMLEAddBin,default)
 S3method(EstMLETriBin,default)
 S3method(Overdispersion,fitBB)
 S3method(Overdispersion,fitGB)
+S3method(Overdispersion,fitGaB)
+S3method(Overdispersion,fitGrIIB)
 S3method(Overdispersion,fitKB)
 S3method(Overdispersion,fitMB)
 S3method(Overdispersion,fitTB)
