### 2019-06-24 c259f74
updated tests for EstMLE functions
