### 2019-06-24 8503cf7
updating tests for methods
