### 2019-06-24 50abbdf
made changes to tests in over dispersion
