### 2019-06-24 #1505 @Amalan-ConStat
As requested I was now added a few more but clear tests focusing on the outputs for the functions pxxxBin,
EstMLExxx and NegLLxxx. Most of these tests are for checking if produce exact values and if their outputs of necessary classes.
