### 2019-06-27 #1505 @Amalan-ConStat
@osorensen  Regarding how i addressed issues pointed out by @jjharden ,
1. I added a dependencies section in the article. Which specifies why the dependency packages are used and what they are.
2. Just an example has been changed to a title called "Modelling", which explains the steps from 1 to 5 in the article more clearly. Where a final conclusion has been reached of which distribution fits the Alcohol data of Week 1.
3. Using the testthat package I have updated the tests which will check if errors are produced, and functions do create necessary outputs as they should.
4. All references have been checked, except two (one article and book) all the others have dois.

In the article highlighted issues have been resolved
1. Binomial to binomial is done everywhere, except when abbreviations are used.
2. has to have is completed in occasions which are necessary.
3.  EstTriBin to EstMLETriBin is completed, which was a spelling miss by me.
4. spaces added as requested in two separate occasions between words.
5. One reference which had a capitalized title has been changed.
