### 2019-06-29 #1505 @csoneson
Thanks @Amalan-ConStat. @jjharden - could you take a look at the revised submission and let us know whether the changes listed above addresses your concerns? Thanks!
