### 2019-06-30 #1505 @jjharden
Everything looks good to me. My comments have been addressed.
