### 2019-06-30 #1505 @csoneson
Thank you @jjharden - could you please also tick the empty boxes in the checklist above?
@Amalan-ConStat - I will take another look at the submission and get back to you shortly to move on with the acceptance procedure.
