### 2019-07-01 #1505 @Amalan-ConStat
I have gone through the article properly and it includes all the corrections without any issues.
