+++ JOSSpaper/Ref.bib
@@ -3,7 +3,7 @@
   title={{Modelling Binary Data}},
   isbn={9780412387906},
   lccn={lc91023845},
-  series={Chapman \& Hall/CRC Texts in Statistical Science},
+  series={{Chapman \& Hall/CRC Texts in Statistical Science}},
   year={1991},
   publisher={Chapman \& Hall},
   doi={10.1002/pst.100}
@@ -142,7 +142,7 @@ doi={10.1007/978-0-8176-4626-4_2}
 @article{COMPoisson,
   title = {{A COM-Poisson type generalization of the binomial distribution and its properties and applications}},
   author={Borges, Patrick and Rodrigues, Josemar and Balakrishnan, Narayanaswamy and Baz{\'a}n, Jorge},
-  journal = {Statistics & Probability Letters},
+  journal = {Statistics \& Probability Letters},
   volume = {87},
   pages = {158 - 166},
   year = {2014},
