### 2019-07-01 #1505 @csoneson
@Amalan-ConStat I have sent a pull request fixing some issues with a couple of the references. If this looks fine to you, please merge it.
The next step is to create a Zenodo archive. Please make sure that the package version number is consistent everywhere, and that the Zenodo archive has exactly the same name and authors as the paper. Then, post the DOI from Zenodo in this thread.
