### 2019-07-01 1010526
version set to 1.4.1 everywhere
