### 2019-07-01 #1505 @Amalan-ConStat
I have completed the merge and it is done. package version number is set to 1.4.1 everywhere, which makes it consistent and accurate. Further, Zenodo archive has exactly the same name and authors as the paper.

http://doi.org/10.5281/zenodo.3265356
