### 2019-07-01 #1505 @csoneson
Hi @Amalan-ConStat - the Zenodo archive in the link above does not have the same title and authors as the paper - the title of the Zenodo archive is "Amalan-ConStat/R-fitODBOD: All issues addressed", and I have been included among the authors.
