### 2019-07-02 #1505 @Amalan-ConStat
Very sorry, I have corrected them now in the zenodo archive. Hope now it is correct. Further article proof is correct.

http://doi.org/10.5281/zenodo.3265356
