### 2019-07-02 #1505 @csoneson
The title is correct now. However, the author names are given in different formats, and consequently the citations of the two records would be different (Mahendran et al. for paper, Amalan et al. for Zenodo archive).
