### 2019-07-02 #1505 @labarba
Big thanks to our editor: @csoneson, and the reviewers: @osorensen, @jjharden — JOSS is thriving thanks to your contributions!
